md5: We don't need to support 64-bit inputs
Saving us 16 bytes if code on x86_64 and 4 bytes of context size.

+++ md5.h
@@ -10,7 +10,7 @@
 #include <stdlib.h>
 
 struct MD5Context {
-    uint64_t size;        // Size of input in bytes
+    uint32_t size;        // Size of input in bytes
     uint32_t buffer[4];   // Current accumulation of hash
     uint8_t input[64];    // Input to be used in the next step
     uint8_t digest[16];   // Result of algorithm
